Relationships

+++ database/migrations/2022_06_10_105205_create_questions_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string("title");
             $table->string("slug");
             $table->text("body")->nullable($value = true)->collation('utf8mb4_general_ci');
-            $table->integer("category_id")->unsigned();
-            $table->integer("user_id")->unsigned();
+            $table->unsignedBigInteger("category_id")->unsigned();
+            $table->unsignedBigInteger("user_id")->unsigned();
             $table->timestamps();
         });
     }
