+++ database/migrations/2022_06_10_105856_create_replies_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('replies', function (Blueprint $table) {
             $table->id();
             $table->text("body")->nullable($value = true)->collation('utf8mb4_general_ci');
-            $table->integer("question_id")->unsigned();
-            $table->integer("user_id")->unsigned();
+            $table->unsignedBigInteger("question_id")->unsigned();
+            $table->unsignedBigInteger("user_id")->unsigned();
             $table->timestamps();
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade')->onUpdate('cascade');
         });
