+++ database/migrations/2022_06_10_110037_create_likes_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->id();
-            $table->integer('reply_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('reply_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
         });
     }
